#头部按钮适配通控切换动画 #修复电源菜单-飞行模式功能无效 #修复组件资源加载错误

+++ app/src/main/java/yunzia/utils/SystemProperties.java
@@ -98,7 +98,7 @@ public class SystemProperties {
     public static int getInt(String key, int def) {
         if (isSupportGetInt) {
             try {
-                return ((Integer) invoke(null, callMethod("getInt", Integer.TYPE), key, Integer.valueOf(def))).intValue();
+                return (Integer) invoke(null, callMethod("getInt", Integer.TYPE), key, def);
             } catch (IllegalArgumentException e) {
                 Log.e(TAG, "key: " + key + " detail:" + e.toString());
             }
